w6_django_project/requirements/base.txt
Django==3.2.6
django-admin-honeypot==1.0.0
djangorestframework==3.6.3
psycopg2==2.7.3
python-decouple==3.1
Pillow==4.2.1
